2.8, 2.9: use `$` and `#` prompts in terminal boxes
I found a way to pick those prompts and put them in a separate,
non-selectable <pre> element on the website version so that you can see
the prompts, but you can also select all the command lines and copy them
without copying the prompts.

It's more instinctive for the typical reader to immediately recognize
command lines when they start with the classic prompts.

Also update CONTRIBUTING.adoc to indicate to use the prompts in
terminal boxes.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>