2.8-2.12: "lost packets" is only meaningful for non-snapshot modes
With a snapshot mode tracing session, the channels are in "overwrite"
mode automatically: the tracers continuously overwrite the oldest
sub-buffers, incrementing their sequence numbers meanwhile.

The overwritten sub-buffers between two snapshots are not lost packets
in this case: they were not meant to be recorded in the first place.

Fix `lttng-docs-2.8.txt` to `lttng-docs-2.12.txt` to clarify this, with
new internal links to tracing session mode descriptions.

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>