Add check and check-VERSION targets
The targets allow the checks to be run for each version (`make
check`), or for a given version (`make check-2.13`).

Multiple checks can be run in parallel with `make check -j$(nproc)`

The pass message output of the check script has been updated to
include the input filename in order to make it clearer which file
succeeds when running checks in parallel.

Change-Id: Ia7aceb97dc35f54ffa791daa285d8bce2dd31ca9